# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AG BACKEND",
        "description": "AG BACKEND",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/apis": {
            "get": {
                "tags": [
                    "APIs"
                ],
                "summary": "List all APIs",
                "operationId": "bc55d3d7c1bfee72242d9f61a7d76c25",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "microservice_id": {
                                                "type": "integer"
                                            },
                                            "route_in": {
                                                "type": "string"
                                            },
                                            "method": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "APIs"
                ],
                "summary": "Create a new API",
                "operationId": "c6350dddde5f4896eb772adbcb830d94",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "microservice_id": {
                                        "type": "integer"
                                    },
                                    "route_in": {
                                        "type": "string"
                                    },
                                    "method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "API created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/apis/{id}": {
            "get": {
                "tags": [
                    "APIs"
                ],
                "summary": "Get API by ID",
                "operationId": "cf07fd2cb930bbc6e94ed9c554b7c5ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "APIs"
                ],
                "summary": "Update an API by ID",
                "operationId": "f65acb2b7dbeeb9de61ab89a33418669",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "microservice_id": {
                                        "type": "integer"
                                    },
                                    "route_in": {
                                        "type": "string"
                                    },
                                    "method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "API updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "APIs"
                ],
                "summary": "Delete an API by ID",
                "operationId": "469cc6cd364f23aa10b59d64e5d88eeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token **_only_**",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}